+++ wn/_queries.py
@@ -9,7 +9,6 @@ import itertools
 import wn
 from wn._types import Metadata
 from wn._db import connect
-from wn._util import split_lexicon_specifier
 
 
 # Local Types
@@ -95,6 +94,7 @@ _ProposedILI = tuple[
     str,  # lexicon
 ]
 _Lexicon = tuple[
+    str,       # specifier
     str,       # id
     str,       # label
     str,       # language
@@ -119,9 +119,9 @@ def resolve_lexicon_specifiers(
         if ':' not in specifier:
             specifier += ':*'
         query = f'''
-            SELECT DISTINCT id || ":" || version
+            SELECT DISTINCT specifier
               FROM lexicons
-             WHERE id || ":" || version GLOB :specifier
+             WHERE specifier GLOB :specifier
                AND (:language ISNULL OR language = :language)
              LIMIT {limit}
         '''
@@ -136,21 +136,20 @@ def resolve_lexicon_specifiers(
 
 
 def get_lexicon(lexicon: str) -> _Lexicon:
-    id, ver = split_lexicon_specifier(lexicon)
     query = '''
-        SELECT DISTINCT id, label, language, email, license,
+        SELECT DISTINCT specifier, id, label, language, email, license,
                         version, url, citation, logo, metadata
         FROM lexicons
-        WHERE id = ? AND version = ?
+        WHERE specifier = ?
     '''
-    row: Optional[_Lexicon] = connect().execute(query, (id, ver)).fetchone()
+    row: Optional[_Lexicon] = connect().execute(query, (lexicon,)).fetchone()
     if row is None:
         raise LookupError(lexicon)  # should we have a WnLookupError?
     return row
 
 
 def get_modified(lexicon: str) -> bool:
-    query = 'SELECT modified FROM lexicons WHERE id || ":" || version = ?'
+    query = 'SELECT modified FROM lexicons WHERE specifier = ?'
     return connect().execute(query, (lexicon,)).fetchone()[0]
 
 
@@ -159,7 +158,7 @@ def get_lexicon_dependencies(lexicon: str) -> list[tuple[str, str, bool]]:
         SELECT provider_id || ":" || provider_version, provider_url, provider_rowid
           FROM lexicon_dependencies
           JOIN lexicons AS lex ON lex.rowid = dependent_rowid
-         WHERE lex.id || ":" || lex.version = ?
+         WHERE lex.specifier = ?
     '''
     return [
         (spec, url, rowid is not None)
@@ -173,11 +172,11 @@ def get_lexicon_extension_bases(lexicon: str, depth: int = -1) -> list[str]:
                (SELECT base_rowid, 1
                   FROM lexicon_extensions
                   JOIN lexicons AS lex ON lex.rowid = extension_rowid
-                 WHERE lex.id || ":" || lex.version = :specifier
+                 WHERE lex.specifier = :specifier
                  UNION SELECT base_rowid, d+1
                          FROM lexicon_extensions
                          JOIN ext ON extension_rowid = x)
-        SELECT baselex.id || ":" || baselex.version
+        SELECT baselex.specifier
           FROM ext
           JOIN lexicons AS baselex ON baselex.rowid = ext.x
          WHERE :depth < 0 OR d <= :depth
@@ -193,11 +192,11 @@ def get_lexicon_extensions(lexicon: str, depth: int = -1) -> list[str]:
                (SELECT extension_rowid, 1
                   FROM lexicon_extensions
                   JOIN lexicons AS lex ON lex.rowid = base_rowid
-                 WHERE lex.id || ":" || lex.version = :specifier
+                 WHERE lex.specifier = :specifier
                  UNION SELECT extension_rowid, d+1
                          FROM lexicon_extensions
                          JOIN ext ON base_rowid = x)
-        SELECT extlex.id || ":" || extlex.version
+        SELECT extlex.specifier
           FROM ext
           JOIN lexicons AS extlex ON extlex.rowid = ext.x
          WHERE :depth < 0 OR d <= :depth
@@ -245,7 +244,7 @@ def find_existing_ilis(
           JOIN ilis AS i ON i.rowid = ss.ili_rowid
           JOIN ili_statuses AS ist ON i.status_rowid = ist.rowid
         '''
-        conditions.append(f'lex.id || ":" || lex.version IN ({_qs(lexicons)})')
+        conditions.append(f'lex.specifier IN ({_qs(lexicons)})')
         params.extend(lexicons)
 
     if conditions:
@@ -259,7 +258,7 @@ def find_proposed_ilis(
     lexicons: Sequence[str] = (),
 ) -> Iterator[_ProposedILI]:
     query = '''
-    SELECT null, "proposed", definition, ss.id, lex.id || ":" || lex.version
+    SELECT null, "proposed", definition, ss.id, lex.specifier
       FROM proposed_ilis
       JOIN synsets AS ss ON ss.rowid = synset_rowid
       JOIN lexicons AS lex ON lex.rowid = ss.lexicon_rowid
@@ -270,7 +269,7 @@ def find_proposed_ilis(
         conditions.append('ss.id = ?')
         params.append(synset_id)
     if lexicons:
-        conditions.append(f'lex.id || ":" || lex.version IN ({_qs(lexicons)})')
+        conditions.append(f'lex.specifier IN ({_qs(lexicons)})')
         params.extend(lexicons)
     if conditions:
         query += ' WHERE ' + '\n           AND '.join(conditions)
@@ -307,7 +306,7 @@ def find_entries(
         conditions.append('e.pos = ?')
         params.append(pos)
     if lexicons:
-        conditions.append(f'lex.id || ":" || lex.version IN ({_qs(lexicons)})')
+        conditions.append(f'lex.specifier IN ({_qs(lexicons)})')
         params.extend(lexicons)
 
     condition = ''
@@ -316,7 +315,7 @@ def find_entries(
 
     query = f'''
           {cte}
-        SELECT DISTINCT e.id, e.pos, lex.id || ":" || lex.version
+        SELECT DISTINCT e.id, e.pos, lex.specifier
           FROM entries AS e
           JOIN lexicons AS lex ON lex.rowid = e.lexicon_rowid
          {condition}
@@ -357,7 +356,7 @@ def find_senses(
         conditions.append('e.pos = ?')
         params.append(pos)
     if lexicons:
-        conditions.append(f'slex.id || ":" || slex.version IN ({_qs(lexicons)})')
+        conditions.append(f'slex.specifier IN ({_qs(lexicons)})')
         params.extend(lexicons)
 
     condition = ''
@@ -366,7 +365,7 @@ def find_senses(
 
     query = f'''
           {cte}
-        SELECT DISTINCT s.id, e.id, ss.id, slex.id || ":" || slex.version
+        SELECT DISTINCT s.id, e.id, ss.id, slex.specifier
           FROM senses AS s
           JOIN entries AS e ON e.rowid = s.entry_rowid
           JOIN synsets AS ss ON ss.rowid = s.synset_rowid
@@ -419,7 +418,7 @@ def find_synsets(
         )
         params.append(ili)
     if lexicons:
-        conditions.append(f'sslex.id || ":" || sslex.version IN ({_qs(lexicons)})')
+        conditions.append(f'sslex.specifier IN ({_qs(lexicons)})')
         params.extend(lexicons)
 
     condition = ''
@@ -433,7 +432,7 @@ def find_synsets(
           {cte}
         SELECT DISTINCT ss.id, ss.pos,
                         (SELECT ilis.id FROM ilis WHERE ilis.rowid=ss.ili_rowid),
-                        sslex.id || ":" || sslex.version
+                        sslex.specifier
           FROM synsets AS ss
           JOIN lexicons AS sslex ON sslex.rowid = ss.lexicon_rowid
           {join}
@@ -452,7 +451,7 @@ def get_entry_forms(id: str, lexicons: Sequence[str]) -> Iterator[_Form]:
           JOIN entries AS e ON e.rowid = entry_rowid
           JOIN lexicons AS lex ON lex.rowid = e.lexicon_rowid
          WHERE e.id = ?
-           AND lex.id || ":" || lex.version IN ({_qs(lexicons)})
+           AND lex.specifier IN ({_qs(lexicons)})
          ORDER BY f.rank
     '''
     pron_query = '''
@@ -476,13 +475,12 @@ def get_synsets_for_ilis(
 ) -> Iterator[_Synset]:
     conn = connect()
     query = f'''
-        SELECT DISTINCT ss.id, ss.pos, ili.id,
-                        sslex.id || ":" || sslex.version
+        SELECT DISTINCT ss.id, ss.pos, ili.id, sslex.specifier
           FROM synsets as ss
           JOIN ilis as ili ON ss.ili_rowid = ili.rowid
           JOIN lexicons AS sslex ON sslex.rowid = ss.lexicon_rowid
          WHERE ili.id IN ({_qs(ilis)})
-           AND sslex.id || ":" || sslex.version IN ({_qs(lexicons)})
+           AND sslex.specifier IN ({_qs(lexicons)})
     '''
     params = *ilis, *lexicons
     result_rows: Iterator[_Synset] = conn.execute(query, params)
@@ -510,22 +508,21 @@ def get_synset_relations(
             (SELECT rowid FROM relation_types {constraint}),
           lexrowids(rowid) AS
             (SELECT rowid FROM lexicons
-              WHERE id || ":" || version IN ({_vs(lexicons)})),
+              WHERE specifier IN ({_vs(lexicons)})),
           srcsynset(rowid) AS
             (SELECT ss.rowid
                FROM synsets AS ss
                JOIN lexicons AS lex ON lex.rowid = ss.lexicon_rowid
               WHERE ss.id = ?
-                AND lex.id || ":" || lex.version = ?),
+                AND lex.specifier = ?),
           matchingrels(rowid) AS
             (SELECT srel.rowid
                FROM synset_relations AS srel
               WHERE srel.source_rowid IN srcsynset
                 AND srel.lexicon_rowid IN lexrowids
                 AND srel.type_rowid IN reltypes)
-        SELECT DISTINCT rt.type, lex.id || ":" || lex.version, srel.metadata,
-                        src.id, tgt.id, tgt.pos, tgtili.id,
-                        tgtlex.id || ":" || tgtlex.version
+        SELECT DISTINCT rt.type, lex.specifier, srel.metadata,
+                        src.id, tgt.id, tgt.pos, tgtili.id, tgtlex.specifier
           FROM matchingrels AS mr
           JOIN synset_relations AS srel ON srel.rowid=mr.rowid
           JOIN relation_types AS rt ON rt.rowid=srel.type_rowid
@@ -558,7 +555,7 @@ def get_expanded_synset_relations(
           reltypes(rowid) AS
             (SELECT rowid FROM relation_types {constraint}),
           lexrowids(rowid) AS
-            (SELECT rowid FROM lexicons WHERE id || ":" || version IN ({_vs(expands)})),
+            (SELECT rowid FROM lexicons WHERE specifier IN ({_vs(expands)})),
           srcsynset(rowid) AS
             (SELECT ss.rowid
                FROM synsets AS ss
@@ -571,9 +568,8 @@ def get_expanded_synset_relations(
               WHERE srel.source_rowid IN srcsynset
                 AND srel.lexicon_rowid IN lexrowids
                 AND srel.type_rowid IN reltypes)
-        SELECT DISTINCT rt.type, lex.id || ":" || lex.version, srel.metadata,
-                        src.id, tgt.id, tgt.pos, tgtili.id,
-                        tgtlex.id || ":" || tgtlex.version
+        SELECT DISTINCT rt.type, lex.specifier, srel.metadata,
+                        src.id, tgt.id, tgt.pos, tgtili.id, tgtlex.specifier
           FROM matchingrels AS mr
           JOIN synset_relations AS srel ON srel.rowid=mr.rowid
           JOIN relation_types AS rt ON rt.rowid=srel.type_rowid
@@ -600,7 +596,7 @@ def get_definitions(
           JOIN synsets AS ss ON ss.rowid = d.synset_rowid
           JOIN lexicons AS lex ON lex.rowid = d.lexicon_rowid
          WHERE ss.id = ?
-           AND lex.id || ":" || lex.version IN ({_qs(lexicons)})
+           AND lex.specifier IN ({_qs(lexicons)})
     '''
     return conn.execute(query, (synset_id, *lexicons)).fetchall()
 
@@ -626,7 +622,7 @@ def get_examples(
           JOIN {table} AS tbl ON tbl.rowid = ex.{prefix}_rowid
           JOIN lexicons AS lex ON lex.rowid = ex.lexicon_rowid
          WHERE tbl.id = ?
-           AND lex.id || ":" || lex.version IN ({_qs(lexicons)})
+           AND lex.specifier IN ({_qs(lexicons)})
     '''
     return conn.execute(query, (id, *lexicons)).fetchall()
 
@@ -651,7 +647,7 @@ def find_syntactic_behaviours(
         conditions.append('sb.id = ?')
         params.append(id)
     if lexicons:
-        conditions.append(f'lex.id || ":" || lex.version IN ({_qs(lexicons)})')
+        conditions.append(f'lex.specifier IN ({_qs(lexicons)})')
         params.extend(lexicons)
     if conditions:
         query += '\n WHERE ' + '\n   AND '.join(conditions)
@@ -675,7 +671,7 @@ def get_syntactic_behaviours(
           JOIN senses AS s ON s.rowid = sbs.sense_rowid
           JOIN lexicons AS lex ON lex.rowid = sb.lexicon_rowid
          WHERE s.id = ?
-           AND lex.id || ":" || lex.version IN ({_qs(lexicons)})
+           AND lex.specifier IN ({_qs(lexicons)})
     '''
     return [row[0] for row in conn.execute(query, (sense_id, *lexicons))]
 
@@ -693,7 +689,7 @@ def _get_senses(
     else:
         raise wn.Error(f'invalid sense source type: {sourcetype}')
     query = f'''
-        SELECT s.id, e.id, ss.id, slex.id || ":" || slex.version
+        SELECT s.id, e.id, ss.id, slex.specifier
           FROM senses AS s
           JOIN entries AS e
             ON e.rowid = s.entry_rowid
@@ -702,7 +698,7 @@ def _get_senses(
           JOIN lexicons AS slex
             ON slex.rowid = s.lexicon_rowid
          WHERE {sourcealias}.id = ?
-           AND slex.id || ":" || slex.version IN ({_qs(lexicons)})
+           AND slex.specifier IN ({_qs(lexicons)})
          ORDER BY s.{sourcetype}_rank
     '''
     return conn.execute(query, (id, *lexicons))
@@ -739,12 +735,11 @@ def get_sense_relations(
           rt(rowid, type) AS
             (SELECT rowid, type FROM relation_types {constraint}),
           lexrowids(rowid) AS
-            (SELECT rowid FROM lexicons WHERE id || ":" || version IN ({_vs(lexicons)}))
+            (SELECT rowid FROM lexicons WHERE specifier IN ({_vs(lexicons)}))
         SELECT DISTINCT rel.type, rel.lexicon, rel.metadata,
-                        s.id, e.id, ss.id,
-                        slex.id || ":" || slex.version
+                        s.id, e.id, ss.id, slex.specifier
           FROM (SELECT rt.type,
-                       lex.id || ":" || lex.version AS lexicon,
+                       lex.specifier AS lexicon,
                        srel.metadata AS metadata,
                        target_rowid
                   FROM sense_relations AS srel
@@ -785,13 +780,13 @@ def get_sense_synset_relations(
           rt(rowid, type) AS
             (SELECT rowid, type FROM relation_types {constraint}),
           lexrowids(rowid) AS
-            (SELECT rowid FROM lexicons WHERE id || ":" || version IN ({_vs(lexicons)}))
+            (SELECT rowid FROM lexicons WHERE specifier IN ({_vs(lexicons)}))
         SELECT DISTINCT rel.type, rel.lexicon, rel.metadata,
                         rel.source_rowid, tgt.id, tgt.pos,
                         (SELECT ilis.id FROM ilis WHERE ilis.rowid = tgt.ili_rowid),
-                        tgtlex.id || ":" || tgtlex.version
+                        tgtlex.specifier
           FROM (SELECT rt.type,
-                       lex.id || ":" || lex.version AS lexicon,
+                       lex.specifier AS lexicon,
                        srel.metadata AS metadata,
                        source_rowid,
                        target_rowid
@@ -840,7 +835,7 @@ def get_metadata(
           FROM {tablename} AS tbl
           JOIN lexicons AS lex ON lex.rowid = lexicon_rowid
          WHERE tbl.id=?
-           AND lex.id || ":" || lex.version = ?
+           AND lex.specifier = ?
     '''
     return connect().execute(query, (id, lexicon)).fetchone()[0] or {}
 
@@ -857,7 +852,7 @@ def get_proposed_ili_metadata(synset: str, lexicon: str) -> Metadata:
           JOIN synsets AS ss ON ss.rowid = synset_rowid
           JOIN lexicons AS lex ON lex.rowid = ss.lexicon_rowid
          WHERE ss.id = ?
-           AND lex.id || ":" || lex.version = ?
+           AND lex.specifier = ?
     '''
     return connect().execute(query, (synset, lexicon)).fetchone()[0] or {}
 
@@ -880,7 +875,7 @@ def get_lexicalized(id: str, lexicon: str, table: str) -> bool:
           FROM {tablename} AS tbl
           JOIN lexicons AS lex ON lex.rowid = tbl.lexicon_rowid
          WHERE tbl.id = ?
-           AND lex.id || ":" || lex.version = ?
+           AND lex.specifier = ?
     '''
     return conn.execute(query, (id, lexicon)).fetchone()[0]
 
@@ -893,7 +888,7 @@ def get_adjposition(sense_id: str, lexicon: str) -> Optional[str]:
           JOIN senses AS s ON s.rowid = sense_rowid
           JOIN lexicons AS lex ON lex.rowid = s.lexicon_rowid
          WHERE s.id = ?
-           AND lex.id || ":" || lex.version = ?
+           AND lex.specifier = ?
     '''
     row = conn.execute(query, (sense_id, lexicon)).fetchone()
     if row:
@@ -909,7 +904,7 @@ def get_sense_counts(sense_id: str, lexicons: Sequence[str]) -> list[_Count]:
           JOIN senses AS s ON s.rowid = c.sense_rowid
           JOIN lexicons AS lex ON lex.rowid = c.lexicon_rowid
          WHERE s.id = ?
-           AND lex.id || ":" || lex.version IN ({_qs(lexicons)})
+           AND lex.specifier IN ({_qs(lexicons)})
     '''
     rows: list[_Count] = conn.execute(
         query, (sense_id, *lexicons)
@@ -925,7 +920,7 @@ def get_lexfile(synset_id: str, lexicon: str) -> Optional[str]:
           JOIN synsets AS ss ON ss.lexfile_rowid = lf.rowid
           JOIN lexicons AS lex on lex.rowid = ss.lexicon_rowid
          WHERE ss.id = ?
-           AND lex.id || ":" || lex.version = ?
+           AND lex.specifier = ?
     '''
     row = conn.execute(query, (synset_id, lexicon)).fetchone()
     if row is not None and row[0] is not None: